refactor: update sveltekit to 1.0.0-next.449
- [breaking] require Node 16.14
- [breaking] Overhaul filesystem-based router
- [breaking] Change load API
