+++ web/.nvmrc
@@ -1,1 +1,1 @@
-v16.10.0
+v16.17.0
