+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Jenkins X UI",
+  "name": "jenkins-x-ui",
   "version": "0.0.1",
   "scripts": {
     "dev": "ROUTE_FOLDER=admin vite dev",
@@ -23,7 +23,7 @@
     "@storybook/addon-svelte-csf": "^2.0.3",
     "@storybook/svelte": "^6.4.22",
     "@sveltejs/adapter-static": "1.0.0-next.38",
-    "@sveltejs/kit": "1.0.0-next.405",
+    "@sveltejs/kit": "1.0.0-next.449",
     "@tailwindcss/forms": "^0.4.0",
     "@tailwindcss/line-clamp": "^0.3.1",
     "@tailwindcss/typography": "^0.5.2",
