+++ web/src/routes/admin/repositories/+page.svelte
@@ -1,30 +1,9 @@
-<script context="module">
-  // Todo: move it to endpoints
-  export async function load({ fetch }) {
-    // ToDo: Use Axios
-    const res = await fetch('http://localhost:9200/api/v1/repositories')
-    const repositories = await res.json()
-    if (res.ok) {
-      const repositoriesProcessed = repositories.map((k) => k.spec)
-      return {
-        props: {
-          repositoriesProcessed,
-        },
-      }
-    }
-    return {
-      status: res.status,
-      // ToDo: use new Error()
-      error: 'Could not fetch repositories',
-    }
-  }
-</script>
-
 <script lang="ts">
   import type { GridOptions } from 'ag-grid-community'
   import Table from '$lib/Components/Table.svelte'
   import { HTMLInColumnCell } from '$lib/cellRenderer'
-  export let repositoriesProcessed
+
+  export let data
 
   let gridOptions: GridOptions = {
     defaultColDef: {
@@ -46,7 +25,7 @@
         },
       },
     ],
-    rowData: repositoriesProcessed,
+    rowData: data.repositoriesProcessed,
     pagination: true,
     paginationAutoPageSize: true,
   }
